% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{get_diffusion}
\alias{get_diffusion}
\title{Obtain the Diffusion operator}
\usage{
get_diffusion(counts, index = NULL, k1 = 10, k2 = 20)
}
\arguments{
\item{counts}{An n by p matrix of counts with cells in rows and genes in columns.}

\item{index}{(Optional:) an index vector for the rows of counts used to compute
local distance.}

\item{k1}{Tuning parameter.}

\item{k2}{Tuning parameter.}
}
\value{
A diffusion object.
}
\description{
Obtain the Diffusion operator
}
